# API Sitira-Sama/penitipanbarang/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/barangDitolaks/{id}": {
      "get": {
        "tags": [
          "barang-ditolak-entity-controller"
        ],
        "description": "get-barangditolak",
        "operationId": "getItemResource-barangditolak-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelBarangDitolak"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "barang-ditolak-entity-controller"
        ],
        "description": "update-barangditolak",
        "operationId": "putItemResource-barangditolak-put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BarangDitolakRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelBarangDitolak"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelBarangDitolak"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "barang-ditolak-entity-controller"
        ],
        "description": "delete-barangditolak",
        "operationId": "deleteItemResource-barangditolak-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "barang-ditolak-entity-controller"
        ],
        "description": "patch-barangditolak",
        "operationId": "patchItemResource-barangditolak-patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BarangDitolakRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelBarangDitolak"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/penitipans/{id}": {
      "get": {
        "tags": [
          "penitipan-entity-controller"
        ],
        "description": "get-penitipan",
        "operationId": "getItemResource-penitipan-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPenitipan"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "penitipan-entity-controller"
        ],
        "description": "update-penitipan",
        "operationId": "putItemResource-penitipan-put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PenitipanRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPenitipan"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPenitipan"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "penitipan-entity-controller"
        ],
        "description": "delete-penitipan",
        "operationId": "deleteItemResource-penitipan-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "penitipan-entity-controller"
        ],
        "description": "patch-penitipan",
        "operationId": "patchItemResource-penitipan-patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PenitipanRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPenitipan"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/satpam/barang": {
      "get": {
        "tags": [
          "penitipan-controller"
        ],
        "operationId": "getAllPenitipan",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Penitipan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "penitipan-controller"
        ],
        "operationId": "addPenitipan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Penitipan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Penitipan"
                }
              }
            }
          }
        }
      }
    },
    "/api/satpam/barang/tolak": {
      "post": {
        "tags": [
          "barang-ditolak-controller"
        ],
        "operationId": "addBarangDitolak",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BarangDitolak"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/BarangDitolak"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/barang": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "getAllPenitipan_1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Penitipan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "addPengajuan_1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pengajuan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/satpam/barang/barang/{id}": {
      "patch": {
        "tags": [
          "penitipan-controller"
        ],
        "operationId": "updateStatusPenitipan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Penitipan"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/barang/{id}": {
      "delete": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "deletePenitipan_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "updateStatusPenitipan_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/satpam/barang/{username}": {
      "get": {
        "tags": [
          "penitipan-controller"
        ],
        "operationId": "getPenitipanByUsername",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Penitipan"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/satpam/barang/tolak/{username}": {
      "get": {
        "tags": [
          "barang-ditolak-controller"
        ],
        "operationId": "getBarangDitolakByUsername",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BarangDitolak"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/satpam/barang/{id}": {
      "delete": {
        "tags": [
          "penitipan-controller"
        ],
        "operationId": "deletePenitipan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/satpam/barang/tolak/{id}": {
      "delete": {
        "tags": [
          "barang-ditolak-controller"
        ],
        "operationId": "deleteBarangDitolak",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth/barang/tolak/{id}": {
      "delete": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "deleteBarangDitolak_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "Pengajuan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "EntityModelPenitipan": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "statuspenitipan": {
            "type": "string"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Penitipan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "statuspenitipan": {
            "type": "string"
          }
        }
      },
...
      "BarangDitolak": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "alasan": {
            "type": "string"
          }
        }
      },
      "EntityModelBarangDitolak": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "alasan": {
            "type": "string"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "BarangDitolakRequestBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "alasan": {
            "type": "string"
          }
        }
      },
...
      "PenitipanRequestBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "barang": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "statuspenitipan": {
            "type": "string"
          }
        }
      },
...
      "StatusRequest": {
        "type": "object",
        "properties": {
          "statuspenitipan": {
            "type": "string"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    }
  }
}