# API Sitira-Sama/penitipanbarang/src/main/java/com/polstat/penitipanbarang/dto/StatusRequest.java
package com.polstat.penitipanbarang.dto;

public class StatusRequest {
    private String statuspenitipan;

    // Getter dan Setter
    public String getStatuspenitipan() {
        return statuspenitipan;
    }

    public void setStatuspenitipan(String statuspenitipan) {
        this.statuspenitipan = statuspenitipan;
    }
}